Add the second menu in the game.

+++ walter2/walter2/headers/menu.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "raylib.h"
 #include "playMenu.h"
-
+#include "sizes.h"
 #define NUMBER_OF_MAIN_BUTTONS   2
 
 
@@ -82,7 +82,7 @@ void menu(void)
         {
             DrawRectangleRec(toggleRecs[i], ((i == mouseHoverRec)) ? MAGENTA : VIOLET);
             DrawRectangleLines(toggleRecs[i].x, toggleRecs[i].y, toggleRecs[i].width, toggleRecs[i].height, PINK);
-            DrawText(processTextMainMenu[i], (toggleRecs[i].x + toggleRecs[i].width / 2 - MeasureText(processTextMainMenu[i] , 10) / 2 ) -25, (int)toggleRecs[i].y + 16, 30, WHITE);
+            DrawText(processTextMainMenu[i], (int)(toggleRecs[i].x + toggleRecs[i].width / 2 - MeasureText(processTextMainMenu[i] , 10) / 2 ) -25, (int)(toggleRecs[i].y) + 16, 30, WHITE);
 
         }
